added delete photo of unused post

+++ src/firebase/deleteFile.ts
@@ -7,6 +7,9 @@ import {
 
 async function deleteFile(imageFileUrl: string) {
     const imageRef = ref(storage, imageFileUrl);
+    if (imageFileUrl === process.env.REACT_APP_DEFAULT_PROFILE_PICTURE) {
+        return new Promise((resolve, reject) => reject("Guest profile picture cannot be deleted"))
+    }
     return deleteObject(imageRef)
         .then(snapshot => {
             return "File deleted"
